+++ backend/Queue/DeobfuscationSteps/3.GetFileInfos/GetFileInfosStep.cs
@@ -63,6 +63,7 @@ public static class GetFileInfosStep
         {
             NzbFile = file.NzbFile,
             FileName = filename,
+            First16KB = file.First16KB,
             ReleaseDate = file.ReleaseDate,
             FileSize = (long?)fileDesc?.FileLength,
             IsRar = file.HasRar4Magic() || file.HasRar5Magic(),
@@ -105,6 +106,7 @@ public static class GetFileInfosStep
     {
         public required NzbFile NzbFile { get; init; }
         public required string FileName { get; init; }
+        public byte[]? First16KB { get; init; }
         public required DateTimeOffset ReleaseDate { get; init; }
         public long? FileSize { get; init; }
         public bool IsRar { get; init; }